fix bugs; delete last entry; more
